% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_truncmvn_cpp}
\alias{r_truncmvn_cpp}
\title{Following Rodriguez-Yam, http://www.stat.columbia.edu/~rdavis/papers/CLR.pdf}
\usage{
r_truncmvn_cpp(x, mu, Sig, B, b)
}
\arguments{
\item{x}{}

\item{mu}{}

\item{Sig}{The covariance matrix}

\item{B}{Transformation matrix for x to give constraints BX <= b}

\item{b}{The vector giving the constraints BX <= b}
}
\description{
Result from Rodriguez-Yam:
 Given
     X ~ N_T(mu,Sigma), where T = {x : Bx <= b}
 Define
     Z := AX, with A s.t. A %*% Sigma %*% t(A) = I (eg. the lower-triangular
                                                    cholesky decomposition)
 Then Z ~ N_S(A%*% mu, I) ~ N(alpha,I)
         where S = {z: Dz <= b}, D = B %*% inverse(A), and alpha = A %*% mu.
}
\details{
Then marginals have structure
   z_j|z_-j ~ N_S_j(alpha_j,1)
         where S_j = {z_j : d_j z_j <= b - D_-j z_-j}
}

