% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rm_sngl_vec_entry}
\alias{rm_sngl_vec_entry}
\title{rm_sngl_vec_entry}
\usage{
rm_sngl_vec_entry(x, jOmit)
}
\arguments{
\item{x}{vector}

\item{jOmit}{The index of the vector x which we'd like to omit from
conditioning and sample)}
}
\description{
Fast conditional sampling multivar. normal vector entry given all others.
}
\details{
\code{rm_sngl_vec_entry}
}

