% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_bayes_cpp}
\alias{rmvn_bayes_cpp}
\title{rmvn_bayes_cpp}
\usage{
rmvn_bayes_cpp(A, b)
}
\arguments{
\item{A}{Precision matrix of the desired sample.}

\item{b}{Vector s.t. mean of desired sample is A^(-1) * b}
}
\description{
Fast sampling from multivar. normal dist'ns for Bayesian hierarchical
models.
}
\details{
\code{rmvn_bayes_cpp} samples from a MV normal of the form
N(A^(-1) * b, A^(-1)).
}

