% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_cond_cpp}
\alias{rmvn_cond_cpp}
\title{Fast conditional sampling multivar. normal vector entry given all others.}
\usage{
rmvn_cond_cpp(mu, cholSigma, iSigma, x, jOmit)
}
\arguments{
\item{mu}{Mean vector of MVN distribution}

\item{cholSigma}{Cholesky decomposition of the covariance matrix of the MVN}

\item{iSigma}{Precision matrix of the MVN}

\item{x}{Vector realization of MVN, with some conponents on which we want
to condition}

\item{jOmit}{The index of the vector x which we'd like to omit from
conditioning and sample)}
}
\description{
\code{r_cond_mvn_cpp2}
}

